// src/types/errorTypes.ts
export interface ErrorClub {
	name?: string;
	establishedAt?: string;
	owner?: string;
	venue?: string;
	manager?: string;
	chairman?: string;
	clubLogoUrl?: string;
	description?: string;
}

export interface ErrorMembers {
	userName?: string;
	firstName?: string;
	lastName?: string;
	dateOfBirth?: string;
	wage?: string;
	country?: string;
	clubId?: string;
	imageUrl?: string;
}